+++ projects/packages/backup/src/js/components/review-request/stories/index.tsx
@@ -11,6 +11,9 @@ const Template = args => <ReviewRequest { ...args } />;
 export const Default = Template.bind( {} );
 Default.args = {
 	cta: 'Text action line, asking for a review',
+	href: 'example-link.com',
 	onClick: action( 'onClick' ),
 	requestReason: 'What triggered the review request (i.e. restore)',
+	dismissedReview: false,
+	onClick: action( 'dismissMessage' ),
 };