+++ projects/packages/search/src/class-options.php
@@ -46,9 +46,16 @@ class Options {
 	 * @var string
 	 */
 	const OVERLAY_TRIGGER_IMMEDIATE = 'immediate';
-	const OVERLAY_TRIGGER_RESULTS   = 'results';
 	const OVERLAY_TRIGGER_SUBMIT    = 'submit';
 	const DEFAULT_OVERLAY_TRIGGER   = self::OVERLAY_TRIGGER_SUBMIT;
+
+	/**
+	 * Deprecated overlay trigger.
+	 *
+	 * @var string
+	 * @deprecated since 11.3
+	 */
+	const OVERLAY_TRIGGER_RESULTS = 'results';
 
 	/**
 	 * Returns a boolean for whether instant search is enabled.