+++ projects/packages/search/src/customizer/class-customizer.php
@@ -128,9 +128,8 @@ class Customizer {
 				'section'     => $section_id,
 				'type'        => 'select',
 				'choices'     => array(
+					Options::OVERLAY_TRIGGER_SUBMIT    => __( 'Open when user submits the form (recommended)', 'jetpack-search-pkg' ),
 					Options::OVERLAY_TRIGGER_IMMEDIATE => __( 'Open when user starts typing', 'jetpack-search-pkg' ),
-					Options::OVERLAY_TRIGGER_RESULTS   => __( 'Open when results are available', 'jetpack-search-pkg' ),
-					Options::OVERLAY_TRIGGER_SUBMIT    => __( 'Open when user submits the form', 'jetpack-search-pkg' ),
 				),
 			)
 		);