+++ projects/plugins/jetpack/extensions/blocks/button/test/edit.js
@@ -20,6 +20,7 @@ const defaultProps = {
 		color: '#000000',
 	},
 	className: 'className',
+	context: {},
 	fallbackBackgroundColor: 'rgba(0, 0, 0, 0)',
 	fallbackTextColor: 'rgba(0, 0, 0, 0)',
 	setAttributes: jest.fn(),
@@ -48,13 +49,6 @@ jest.mock( '@wordpress/block-editor', () => ( {
 		gradientValue: undefined,
 		setGradient: jest.fn(),
 	} ),
-} ) );
-
-// Temporarily mock out the ButtonWidthControl, which is causing errors due to missing
-// dependencies in the jest test runner.
-jest.mock( '../button-width-panel', () => ( {
-	...jest.requireActual( '../button-width-panel' ),
-	ButtonWidthControl: () => <div>Mocked Width Control</div>,
 } ) );
 
 beforeEach( () => {