+++ projects/plugins/jetpack/extensions/blocks/subscriptions/controls.js
@@ -10,7 +10,7 @@ import { ToggleControl, PanelBody, RangeControl, TextareaControl } from '@wordpr
 import { createInterpolateElement } from '@wordpress/element';
 import { __, _n, sprintf } from '@wordpress/i18n';
 import InspectorNotice from '../../shared/components/inspector-notice';
-import { ButtonWidthControl } from '../button/button-width-panel';
+import { WidthControl } from '../../shared/width-panel';
 import {
 	MIN_BORDER_RADIUS_VALUE,
 	MAX_BORDER_RADIUS_VALUE,
@@ -209,7 +209,7 @@ export default function SubscriptionControls( {
 					onChange={ newSpacingValue => setAttributes( { spacing: newSpacingValue } ) }
 				/>
 
-				<ButtonWidthControl
+				<WidthControl
 					width={ buttonWidth }
 					onChange={ newButtonWidth => setAttributes( { buttonWidth: newButtonWidth } ) }
 				/>